ad_genius_clean_upload/requirements.txt
google-adk>=0.4.0
google-cloud-aiplatform
google-genai
requests
beautifulsoup4
pypdf
tenacity
python-dotenv
pandas
